Add export support for ticket search page
Adds a link to the scp/tickets.php page (Tickets) allowing for the export of
the data to CSV file. This is preliminary and still needs a few items:
  * An icon + button CSS definition
  * Support for alternate columns (staff assigned, etc.)

Also add beginnings of complex export support (PDF, JSON  and other
formats
to follow). CSV is fully supported for now.

And for brevity, add a Http::download method to allow for consistent and
browser-independent forced page content downloads