Add annulment support for ticket events
Add the ability of revoking previous ticket state tracking events when new
events are logged for the same ticket. This will allow, for instance, the
ability to revert the 'closed' state of a ticket when the ticket is
reopened.

For statistics tracking, a user could configure whether or not the events
should be counted for each event tracked or just the non-annulled events.
For instance, if a ticket is closed and reopened several times, only the
very last closed event should count toward the statistics for the ticket.
Therefore, when a ticket is reopened, previous closed events should be
marked as annulled.