Handle tickets assigned to teams
Add Team::hasMember utility method to assist in determining if a ticket is
owned by a team that he currently-logged-in-staff is a member of. Then,
check if the ticket is assigned to a staff member other than the current
staff or to a team that the current staff is not a member of. If so (the
ticket is not assigned to the current staff), then display the warning
banner showing who currently owns the ticket.