Live migrate the PHP session to database
Migrate the PHP session from disk to database live. To pull this off, the
session contents are written to database under the current session id. When
the `ostversion` column is dropped from the %config table, the system will
automatically switch to database-backed sessions in osTicket 1.7 mode.

This is sort-of hacked together by carefully calling a instance method of
the osTicketSession class statically, and modify the instance method to
support static invocation.