Fix AttachmentMigrater::queueAttachments
Remove clobber between queue() method and queue instance property (which is
the queue list). Also properly limit the number of attachments queued for
migration as indicated by the initial call to the queueAttachments() method.
Lastly, fix an issue where the first $limit items of the database cannot be
found on or read from the disk. Recurse through the queueAttachments()
method until the entire database is scanned or $limit items are placed in
the queue.