Add UPGRADING.txt file to describe post automated upgrade activities
The upgrader can't (and shouldn't) do everything necessary to complete an
upgrade. The adminsitrator will need to take care of a few extra tasks that
are outlined in the UPGRADING.txt file. One of the items includes removing
source code no longer used in the 1.7 codebase. A script is shipped (Unix
only) that allows an administrator to automatically clean up all old source
files that are no longer used.