Fix md5 hash on windows platforms
The installer compares the md5 of the on-disk sql install file with the
signature file shipped alongside. If the signatures do not match, it will
refuse to install osTicket. On Windows(TM) platforms, PHP will automatically
translate newlines to Windows newlines (CRLF), which will corrupt the MD5
hash of the file.

This patch opens the file in binary mode explicitly to prevent the automatic
conversion.