Add deploy cli script
This will allow folks following the git development model to checkout the
sourcecode to one folder, follow an upstream git brach, and deploy the
osticket installation from a current branch to a location somewhere.

The deploy process differs from the unpack process in that it does not
assume to be inside an 'upload/' folder from the install zipball, and it
does not deploy the 'setup/' folder.

The include folder is still supported to be installed elsewhere.