Use a failsafe encoding for unrecognized charsets
If the body of an email message is written and encoded with an unrecognized
charset (like 'iso', for instance), then use the original text, assumed to
be 8-bit encoded. This patch will create idiosyncrasies, where text might be
transcoded to utf-8 incorrectly, but it should eliminate instances where
email message bodies are missing due to incorrect charset labels.