Implement transient SLAs
SLAs can be marked transient. When a ticket is assigned to a transient SLA
and it is transferred to a department or help topic having a default SLA,
the SLA will change to the SLA of the new department or help topic. This
process can continue as long as the ticket has a transient SLA assigned.
Once a non-transient SLA is assigned to the ticket, the SLA will no longer
change automatically. Thereafter, the SLA can only be manually changed.