Make phone ext optional for XML API
Previously, the @ext attribute for the phone number field was assumed to be
included. An XML payload without the @ext attribute would have been
corrupted to include only the first digit of the phone number, which would
fail validation.

This patch correctly handles the XML payload without the @ext attribute for
the phone number. Subsequently, the <phone_ext> element is now also valid
(and optional as well).

Fixes #670