Provide delay on inline-image hover
This means that drive-by mouse hovers, where the user likely did not intend
for the hover to actually popup the banner with the filename and download
button, do not trigger the popup. This patch changes the popup effect so
that it requires a 500ms hover before the popup is shown.