Continue drafts after a reset
Previously, deleting a draft either from a form [reset] button or from the
trashcan would delete the draft and cancel the draft system for the
remainder of the page's lifetime. This patch deletes the current draft and
resets the Redactor's content, but then it starts a new draft cycle so
drafts can be continued. It also prevents posting to an undefined URL if the
draft_id has not yet been setup.

Fixes #90