Fix filter issue where negative logic prevents matches
The quickList() method will attempt to ask the database to find filters that
might match the incoming ticket information. The idea is that MySQL is
likely faster than PHP. The problem is that it assumes positive logic is
being utilized.

This patch adds all filters with at least one rule with dn_contain
(does-not-contain) or not_equal (not-equal)