Handle email addresses used as names
If an email arrives in the system without a "personal" field (name), the
email address is used as the name. The email address, however, is not parsed
well by the PersonsName constructor. This patch uses the mailbox of the
email address as the name.

Also include a slight boost to the mbstring compatibility functions declared
in the bootstrap::i18n_prep() function