qr: Add concept of top-level hash algorithm
The SmartFuzzyHash will be used to capture multiple hashes for an email with
alternative bodies, for instance, and will match an incoming text against
any of the saved hashes. It will also scrub the incoming text against
several scrubbing vectors in an attempt to get the text to match closer to
the original and have more of its content removed.

Also remove : delimiters in the hash output as str_split() can recreate the
hash sections without delimiters.