Fixup querying drop-down lists via advanced search
It turns out that the _id field was not added to the cdata table when it was
created for SelectionField items, because the ::hasIdValue() method was not
implemented. This patch correctly searches SelectionFields via advanced
search and includes a patch to drop the cdata table if it exists (which will
trigger it to be rebuilt when needed).