Allow for duplicate tickets
osTicket uses message-id to detect previously "seen" emails in order to
prevent duplicate tickets from being created from a single email. This is
necessary to avoid duplicate tickets when fetched emails are left in the
INBOX.

The changes allows for less strict matching which can allow for
duplicate tickets when an email is sent to multiple email queues.

Scenario:
* We have multiple departments with strict access control on tickets
  e.g members of Billing department do not have access to tickets in Legal department
* Joe sends an email to billing@company.com and legal@company.com (CC)
* If a ticket gets created in Legal first (unpredictable - depends on which inbox gets processed first)
  - Billing department will be left in the dark
  - No way for legal to know that the same email was also addressed to Billing