thread: Detect inline and separate attachments
If an email contains both inline and separate attachments, the previous
logic would detect all the attachments as inline.

Separate attachments have the `cid` set to `false`. This will pass the
`isset` test, which would cause the logic to search through the body for a
string of `cid:`, which would very likely be found if there was another
inline image somewhere in the body of the email.