forms: Add concept of inline forms
In-line forms are forms that are rendered as one field. The data of the
inline form is also saved in the data for one field. Currently, the data is
rendered to JSON and stashed in the database.

The data in the field is also accessible via the variable replacement
system, so something line %{ticket.field.subfield} can be handled by the
inline form field.