Actually require "required" fields to not be blank
When closing a ticket which uses custom required internal fields, the
required fields are not actually required if you never edit the ticket.
This update forces you to fill out the required fields before closing
the ticket. It's necessary because gathering custom data on the tickets
(customer name for example) is pointless when people are closing them
without adding this info.