Implement concept of custom ticket numbers
Ticket numbers are now configurable as a system-default and by help topic.
Previously, an administrator could select between six-digit random and
sequential numbers. Now, the sequence is selectable so multiple sequences
as needed, and a format can be specified to define how the ticket numbers
look. Formats can also be defined for each help topic.