Disable closing tickets with empty required fields
These modifications force required fields to be non-empty before closing
a ticket. Previously, tickets could be created, worked, and closed
without entering data into custom required fields (so long as the
custom-fields were internal and the ticket was never edited internally)
It does this by disabling the "close on reply" checkbox and displaying a
warning on the close button popup.