i18n: Use APC to stash the translation file
Turns out including a PHP file in a PHAR file is not cacheable by APC. This
patch adds APC stash support to the translation loading phase, if supported.
This seems to save about 2.5ms (3.6ms down to 1.1ms) processing time per
request. Most of the savings is in the PHAR file processing time (signature
verification), which is avoided if the translation can be loaded from APC.