Make date searching more intuitive
If you entered a date search range from 2014-01-01 to 2014-01-02 you
would only get results from 2014-01-01 because $endtime is set to
midnight. For the same reason, if you search for tickets from 2014-01-01
to 2014-01-01 you would never get any results. That can be confusing for
the end user. Intuition might say just add 86400 to the $endtime
(seconds in 24 hours). But then I realized not all days have 24 hours...