forms: Add new visibility property
This allows fields to specify a visibility constraint that will be evaluated
in real time in the browser to automatically show and hide fields that
should be hidden based on values of other fields in the same form.

Validation is not performed on fields server-side if they are considered
invisible when submitted.