Make ticket subject better visible
The idea for this pull request is based on the suggestion of mLipok from the forum:
http://osticket.com/forum/discussion/79243/ticket-subcject-did-not-catches-the-eye#latest

We also found that sometimes the ticket subject / issue summary could be better visible. So I played a bit around and now that is the results. Move the subject next to the ticket thread, increased font size from 11pt to 12pt and decreased padding from 10px to 8px. It's just a little cosmetic change, but I (and also my colleagues) think that it's better than before. If you like it, merge it, if not I will delete it.