forms: Support non-US locale on client portal
The staff portal has a feature where the date fields are converted to
YYYY-MM-DD prior to form submission. However, this code was not ported to
the client portal. Therefore, for non-US date formats (d/m/Y for instance)
would not be processed properly by the server and would be nulled prior to
validation.