Added Base time type & fields
I have updated the database and code using the information provided by
Edge to add a time type to the tickets. The next step is to replicate
the code for ticket statuses allowing users to create their own time
types.

I have also created a core stream called 731b6166.patch.sql which
contains the database modifications required for this to work.