Separate 'required to open' and 'required to close'
The custom-field requiredness is now split into "required for opening a
ticket" and "required to close a ticket". This allows for opening a
ticket when all the information which is required for closing the ticket
it not yet known. A new `close_required` field needs to be added to the
database FORM_FIELD_TABLE to make this work--

ALTER TABLE `FORM_FIELD_TABLE` ADD `close_required` BOOL NOT NULL AFTER
`required`