i18n: Let the browser truncate long texts
This helps in long UTF-8 characters as the bytes were previously counted and
the width of the characters was assumed to correlate with the count of the
bytes in the text. This method allows the browser to calmly truncate the
text according to the available space in the page where the text is
rendered.