oops: Fix empty export in some cases
If the ticket details form has an extra field of type "Priority Level"
(beyond the one built in), exports of tickets will be empty.

This patch addresses the issue in the exporter which uses an older version
of the custom data materialized view (__cdata table), which created separate
columns for selection and ID values. The current cdata system only creates a
single column and stores the ID value. It also addresses an issue where the
ID column was passed to the PriorityField::to_php() as the first argument.