Notification of overdue tickets
Set up the system to provide mechanisms to send periodically a reminder
of overdue tickets.
Two methods are added to the Ticket class. One is notifyOverdue that
will send the message and the second is triggerOverdue that will find
the overdue tickets in the system.
Finally notify_overdue.php has been added in the api folder. This can
be run periodically with crontab.