Implement resend and arbitrary history
Add in the ability for an agent to resend a response. Optionally editing the
response before sending it, and setting the signature as is possible with
the usual responses.

When the response is resent, the edited version is marked as GUARDED, and
subsequent edits will result in new links in the history chain. That is,
when a response is edited and resent by an agent, that response will remain
in the history chain.