email: Some bounce notices actually have rich content
Here's an example email structure:
```
multipart/mixed
 - multipart/report; delivery-status
   - multipart/alternative
     - text/plain
     - text/html
   - message/delivery-status
   - message/rfc822
```

The previous code would only find the body if the email main headers had:
Content-Type: multipart/report; report-type="delivery-status". In such a
case it would scan for a plain/text body.

This patch will scan for the usual body if the scan for the body as usual if
the report scan did not find anything.

Also, output errors to stderr when running API from the command line