mail: Ensure proper encoding of the To header
When using the php mail() function, the To: header encoded in the Mail_mime
library was not properly transferred to the $to variable passed to the mail
backend. The mail() function will add its own To: header which will
mimic the received destination address.

This patch ensures that the $to parameter is properly encoded.