file: Allow different names for duplicated content
This patch (which needs further development), introduces an extra field to
the %attachment table to store a file's name, in the event that two files
with the same content are stored in the database with differing names. In
such a case, the name is stored in the %attachment table. This allows the
attachment to specify a different name each time the same file is attached
to something in the system.