variable: Rework the variable replacement system
getVar() is now called first, then object properties and array properties
are considered, and then getTag() is considered if none of the other options
work. This allows for more predictable results using the variable
replacement system, and allows ::getVar() to override every other means to
use variables in email templates.