typeahead: Implement `delay` option
This will allow the typeahead to wait a short amount of time before fetching
results from the backend, which will allow for request debouncing per se. In
other words, not every keystroke will signal a request to the backend to
update the results. Instead, when the user slows down typing, the request
will begin thereafter.