upgrade: Defer migrating attachments from v1.6
This patch suggests that old attachments in the `upload_dir` can be located
and left alone when upgrading away from osTicket v1.6. Later, via the `file`
command line applet, the attachments can be migrated into the database, or
to the new filesystem plugin after it is installed with:

    php manage.php file migrate --backend=6 --to=D

This allows attachment migration to be retried in the event it failed for
any reason.