upgrade: Consolidate updates to %thread_entry
The previous code resulted in rebuilding the %thread_entry table several
times during the upgrade process. For systems will several hundred or
thousand megabytes of content in their system, this could result in a very
slow or crashed system.

This patch avoids multiple rebuilds of the table by simply creating a new
one and copying the correct values into the new table. It also avoids
possible duplicates when changing collaborators to the new thread system
considering that collaborators were not properly removed with tickets in
previous versions.