Add option to use dept/agent name on replies
Enable the ability to show the agent's full name or department's name on
responses/replies in addition to the emails address name. Email address of
the response would still be designated outgoing email address of the
Department to which the ticket is assigned or the system default outgoing
email address.

* Agents can set the default FROM Name in Profile > Preference.
* Private department's name won't be used regardless of the setting.
* Agent's name won't be used if identity masking is enabled.