queue: Reuse column definitions
This allows for less work when designing new queues. The column design—the
data definition, annotations, and conditions, are saved with the column. The
heading, sort order, and width are set when the column is associated with a
queue. This means that changing the behavior of a particular column will
mean not having to redo work for each queue that uses the column.