UI Utility navigation and profile block.
UI Prototype changes the layout and function of the top right corner
navigation. Creates a profile panel with information related to the
Agent and a link to the Agents full profile. Also moves the Agent,
Admin, and Account navigation into a drop down with the possibility for
future expansion to the navigation. i.e. Report a bug that allows
admins to submit tickets directly to the osTicket Team. Or access help
information. Ideally admins could add their own custom navigation under
the primary links.

TO-DO:
1. needs Php, it’s only HTML and CSS right now.