thread: Add `started` event
This is used to differentiate between when a task is created and when the
task is started. It will help explain the timeline gaps for dependent canned
tasks, where several tasks are created at the same time but are started in
some hierarchical order rather than simultaneously.

It also adds database support for an `other` event type, which may help in
adding new event types for extensions without the need for special database
modifications.