Added Banned identification to "User Directory"
Added a easy to see "Banned)" in red nest to the Guest text in the
Status column within the User Directory to make it easier to see if a
user or client is banned from sending emails into the ticket system.
This has been added in the User Directory List as well as when you view
individual user details.